# Implement simple caching for -fstype