--- a/tests/find_cmd_tests.rs
+++ b/tests/find_cmd_tests.rs
@@ -785,10 +785,12 @@ fn find_age_range() {
 #[serial(working_dir)]
 fn find_fs() {
     use findutils::find::matchers::fs::get_file_system_type;
+    use std::cell::RefCell;
     use std::path::Path;
 
     let path = Path::new("./test_data/simple/subdir");
-    let target_fs_type = get_file_system_type(path).unwrap();
+    let empty_cache = RefCell::new(None);
+    let target_fs_type = get_file_system_type(path, &empty_cache).unwrap();
 
     // match fs type
     Command::cargo_bin("find")