--- a/tests/find_cmd_tests.rs
+++ b/tests/find_cmd_tests.rs
@@ -830,6 +830,19 @@ fn find_fs() {
         .success()
         .stdout(predicate::str::is_empty())
         .stderr(predicate::str::is_empty());
+
+    let path = Path::new("./test_data/links");
+    let empty_cache = RefCell::new(None);
+    let target_fs_type = get_file_system_type(path, &empty_cache).unwrap();
+
+    // working with broken links
+    Command::cargo_bin("find")
+        .expect("found binary")
+        .args(["./test_data/links", "-fstype", &target_fs_type])
+        .assert()
+        .success()
+        .stdout(predicate::str::contains("./test_data/links/link-missing"))
+        .stderr(predicate::str::is_empty());
 }
 
 #[test]