Fix #418.

Before:

```
Benchmark 1: find ~/Workspace -fstype xfs
  Time (mean ± σ):     195.8 ms ±   3.2 ms    [User: 34.7 ms, System: 160.7 ms]
  Range (min … max):   192.4 ms … 202.4 ms    14 runs

Benchmark 2: target/release/find ~/Workspace -fstype xfs
  Time (mean ± σ):     12.494 s ±  0.098 s    [User: 2.645 s, System: 8.936 s]
  Range (min … max):   12.318 s … 12.670 s    10 runs

Summary
  find ~/Workspace -fstype xfs ran
   63.80 ± 1.16 times faster than target/release/find ~/Workspace -fstype xfs
```

After:

```
Benchmark 1: find ~/Workspace -fstype xfs
  Time (mean ± σ):     190.7 ms ±   3.4 ms    [User: 30.6 ms, System: 159.8 ms]
  Range (min … max):   186.9 ms … 198.0 ms    15 runs

Benchmark 2: target/release/find ~/Workspace -fstype xfs
  Time (mean ± σ):     284.1 ms ±   3.2 ms    [User: 67.8 ms, System: 215.7 ms]
  Range (min … max):   278.7 ms … 289.0 ms    10 runs

Summary
  find ~/Workspace -fstype xfs ran
    1.49 ± 0.03 times faster than target/release/find ~/Workspace -fstype xfs
```

Not exactly as fast, but significantly faster than no caching.